requirements.txt
Django==2.0.4
django-cors-headers
djangorestframework
PyMySQL
Pillow
pyjwt
pycryptodome
selenium
qiniu